Use c++ class for fixed interpolator
